requirements.txt
flask
glunicorn
google-generativeai